==> CBF_code/01_data_preprocess.ipynb <==
Processing the business dataset, to find US popular open restaurants.

import pandas as pd
df_yelp_business = pd.read_json('yelp_academic_dataset_business.json', lines=True)
df_yelp_business

# categories column has many categories at once. Let's split the categories column and store each category in it's own column
business_category_split = df_yelp_business['categories'].str.split(',', expand=True)[[0,1,2]]
# rename the new columns
business_category_split.columns = ['category_1', 'category_2', 'category_3']
df_yelp_business = pd.concat([df_yelp_business, business_category_split], axis=1)
# now drop the actual categories column
df_yelp_business = df_yelp_business.drop(['categories'], axis=1)
# print the head of the dataset
df_yelp_business.head()

df_yelp_business['city']

# let's filter our dataset.
# we'll use Restaurants category
df_yelp_restaurants = df_yelp_business.loc[(df_yelp_business['category_1']=='Restaurants') | (df_yelp_business['category_2']=='Restaurants') | (df_yelp_business['category_3']=='Restaurants')]
df_yelp_restaurants

city_cnt = df_yelp_restaurants['city'].value_counts()
print(city_cnt)
state_cnt = df_yelp_restaurants['state'].value_counts()
print(state_cnt)

# df_us_restaurants = df_yelp_restaurants.drop(df_yelp_restaurants[df_yelp_restaurants['state'] == 'AB'].index)
# df_us_restaurants

open_cnt = df_yelp_restaurants['is_open'].value_counts()
print(open_cnt)
# us_restaurants = df_yelp_restaurants.drop(df_yelp_restaurants[df_yelp_restaurants['is_open'] == 0].index)
# us_restaurants

print('Mean of restaurants review_count is: ', df_yelp_restaurants['review_count'].mean())
print('Median of restaurants review_count is: ', df_yelp_restaurants['review_count'].median())
print(df_yelp_restaurants['review_count'].value_counts())

pop_rests = df_yelp_restaurants.drop(df_yelp_restaurants[df_yelp_restaurants['review_count'] <= 20].index)
pop_rests.to_csv("pop_rest.csv", index = None)

# pop_rests['state'].value_counts()

Process the review dataset, only keep reviews for our target restaurants.

import json
review_file = open("yelp_academic_dataset_review.json", encoding='utf-8')
review = []
for line in review_file:
    review.append(json.loads(line))

df_yelp_reviews = pd.DataFrame(review)
review_file.close()
df_yelp_reviews

pop_reviews = df_yelp_reviews[df_yelp_reviews['business_id'].isin(pop_rests['business_id'])]
pop_reviews

pop_reviews.to_csv("pop_reviews.csv", index=None)

import pandas as pd
reviews_df = pd.read_csv('pop_reviews.csv')
df_res_bus = pd.read_csv('pop_rest.csv')
df_merge = pd.merge(reviews_df, df_res_bus[['city', 'business_id',"review_count"]], how = "outer", on ="business_id")
df_merge

import json
import pandas as pd
user_file = open("yelp_academic_dataset_user.json", encoding='utf-8')
user = []
for line in user_file:
    user.append(json.loads(line))

df_yelp_users = pd.DataFrame(user)
user_file.close()
df_yelp_users

users_ids_subset = df_merge['user_id'].unique().tolist()
user_data_samp = df_yelp_users[df_yelp_users['user_id'].isin(users_ids_subset)]
user_data_samp

user_data_samp.to_csv("pop_user.csv", index=None)

unique_user = pop_reviews['user_id'].nunique()
print('unique user in review data: ', unique_user)
unique_rest = pop_reviews['business_id'].nunique()
print('unique restaurant in review data: ', unique_rest)

sum_text = pop_reviews['text'].describe()
print(sum_text)
review_nan = pop_reviews.isnull().sum()
print(review_nan)

sum_stars = pop_reviews['stars'].describe()
print(sum_stars)
print(type(pop_reviews['stars']))

Create the user-matrix, where rows indicates users and columns indicates restaurants.
