==> CBF_code/02_review_filtering.ipynb <==
import pandas as pd
import numpy as np

## Preview data

busi = pd.read_csv('pop_rest.csv')
rev = pd.read_csv('pop_reviews.csv', parse_dates=['date'])
user = pd.read_csv('pop_user.csv', low_memory=False)

# business_data
busi.head(1)

# review_data
rev.head(1)

# user_data
user.head(1)

## Concat review_df & business_df

rev_ = rev[['review_id', 'user_id', 'business_id', 'stars', 'text', 'date']]
busi_ = busi[['business_id', 'name', 'city', 'state', 'category_1', 'category_2', 'category_3']]

# add business info onto each review
rev_busi_ = rev_.merge(busi_, how='outer', on='business_id')
rev_busi_ = rev_busi_.reset_index(drop=True)
rev_busi_

rev_busi_.describe() # 1123142 reviews

## Filter reviews: only keep the latest review for each user on each business

user_busi_groupby = rev_busi_.groupby(['user_id', 'business_id'])
review_id_lst = []
for group_name, group_df in user_busi_groupby:
    group_df.sort_values(by='date')
    group_df.reset_index(inplace=True, drop=True)
    review_id_lst.append(group_df.iloc[-1,0]) # the latest review by each user on each business

rev_busi_ = rev_busi_[rev_busi_['review_id'].isin(review_id_lst)]
len(rev_busi_) 

## Filter users & business based on reviews' counts

cnt_df = rev_busi_[['user_id', 'review_id']].groupby('user_id').count()
cnt_df['review_id'].mean()
cnt_df['review_id'].median()

def review_filters(data, filtered_var, base_var, threshold):
    '''
    only keep the values of filter_var that meets:
    filter_var[base_var].sum() > threshold
    '''
    count_df = data[[filtered_var, base_var]].groupby(filtered_var).count()
    valid_id = count_df[count_df[base_var]>threshold].index.tolist()
    data = data[rev_busi_[filtered_var].isin(valid_id)]
    return data

# user_id['review_id'].sum() > 5
rev_busi = review_filters(rev_busi_, 'user_id', 'review_id', 5)

# business_id['review_id'].sum() > 20
rev_busi = review_filters(rev_busi, 'business_id', 'review_id', 20)

rev_busi.describe() 

## Output the filtered csv

rev_busi.to_csv('filtered_reviews.csv')
