==> CBF_code/03_feature_engineer_business_attr.ipynb <==
# Feature Engineering - Business Attributes

import pandas as pd
import numpy as np
from sklearn.preprocessing import StandardScaler

rev_busi= pd.read_csv('filtered_reviews.csv', parse_dates=["date"])
busi = pd.read_csv('pop_rest.csv')
busi

### Change dict attribute to dummy variables

def convert_dict_into_dummy(data,feature):
    """
    First change feature values from str to dict, then create variables according to dict keys.
    return: dataframe with dict keys as columns
    """
    col_index = data.columns.get_loc(feature)
    get_dict = pd.Series(data.iloc[:,col_index].replace(np.nan,"None")).apply(eval)
    dummy_df = get_dict.replace("None",np.nan).replace("nan",np.nan).apply(pd.Series)
    return dummy_df

attr = convert_dict_into_dummy(busi,"attributes")
attr.head(1)

attr.columns.to_list()

ambience = convert_dict_into_dummy(attr,"Ambience")
ambience.dropna().head(1)

# change column name
ambience.columns = ["Ambience_"+i for i in ambience.columns.tolist()]
# concat
attr = attr.drop('Ambience', axis = 1)
attr = pd.concat([attr,ambience],axis=1)

attr

### GoodforMeal, latenights

goodformeal = convert_dict_into_dummy(attr,"GoodForMeal")
goodformeal.dropna()

### Change dict into boolean

def convert_dict_into_boolean(data,feature,new_name):
    """
    For some features that have many nan, but still have several values,
    convert it into boolean.
    """
    col_index = data.columns.get_loc(feature)
    data[new_name] = False
    for i in range(len(data)):
        if pd.isna(data.iloc[i,col_index]):
            continue
        elif "True" in  data.iloc[i,col_index]:
            data.loc[i,new_name] = True
    
    return data

attr = convert_dict_into_boolean(attr,"BusinessParking","Parking")
attr = attr.drop("BusinessParking",axis=1)

attr = convert_dict_into_boolean(attr,"Music","music")
attr = attr.drop("Music",axis=1)

### Hours

hours = convert_dict_into_dummy(busi,"hours")
hours.notnull().head()

###  Concatenate to form final business features

# busi = busi.drop
bus_df = pd.concat([busi.drop(['attributes','hours'],axis=1),attr,hours.notnull()],axis=1)
bus_df.head(1)

### Data cleaning

def delete_u(data,feature):
    col_index = data.columns.get_loc(feature)
    values = data.iloc[:,col_index].value_counts().index
#     print(values)
    for i in values:
        if i == "None":
            data.iloc[:,col_index].replace("None",np.nan,inplace=True)
        else:
            data.iloc[:,col_index].replace(i,i.split("'")[1],inplace=True)
            # for Alcohol
            data.iloc[:,col_index].replace("none",np.nan,inplace=True)
    return data      

for feature in ["RestaurantsAttire","Alcohol","NoiseLevel","Smoking","WiFi"]:
#     print(feature)
    bus_df = delete_u(bus_df,feature)

bus_df.head(1)

bus_df["RestaurantsAttire"].value_counts()

### Drop non-related columns

bus_df.columns.to_list()

bus_df = bus_df.drop(["DietaryRestrictions", "BYOB", "GoodForMeal", "AgesAllowed","Open24Hours","AcceptsInsurance",
                      "BYOBCorkage"],axis=1)

bus_df = bus_df.replace('True',True)
bus_df = bus_df.replace('False',False)
bus_df = bus_df.replace('None', np.nan)
bus_df = bus_df.replace('nan', np.nan)
bus_df.shape

bus_df_subset = bus_df[bus_df.business_id.isin(rev_busi["business_id"].unique())]
bus_df_subset.shape

features_ind = bus_df_subset.columns.get_loc("RestaurantsGoodForGroups")
features = bus_df_subset.columns[features_ind:]
features
bus_df_subset = bus_df_subset.set_index("business_id").filter(features)

bus_df_subset.head(1)

### Impute Missing Values

bus_df_subset = bus_df_subset.fillna(False)
bus_df_subset = bus_df_subset * 1

### Correct object data types:

# feature_dtypes = []
# for i in features:
#     print(i)
#     type_to_convert = type(bus_df_subset[i].iloc[0])
#     print(type_to_convert)
#     bus_df_subset[i] = bus_df_subset[i].astype(type_to_convert)

## Drop columns that remain objects
col_index = bus_df_subset.columns[bus_df_subset.dtypes != "object"]
bus_df_subset = bus_df_subset[col_index]

bus_df_subset = busi[["business_id","latitude", "longitude", "stars", "review_count", "is_open"]].set_index("business_id").merge(\
                      bus_df_subset,  left_index = True, right_index = True )  

###  Standardize non-boolean variables

scaler = StandardScaler()
vars_to_scale = ["latitude", "longitude", "stars","review_count"]
bus_df_subset[vars_to_scale] = scaler.fit_transform(bus_df_subset[vars_to_scale])
bus_df_subset.to_csv("business_subset_cleaned.csv")

bus_df_subset
